# specs/001-invoice-service/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Invoice Management Service API
  version: 1.0.0
  description: |
    Comprehensive Invoice Management Service for MALIEV Co. Ltd.

    This service serves as the single source of truth for all invoice-related data, handling:
    - Invoice creation (from quotations or manual entry)
    - Invoice splitting and reconciliation
    - Payment tracking and allocation
    - Currency conversion with exchange rate persistence
    - Comprehensive audit trails with 7-year retention
    - Integration with Currency, Quotation, PDF, and Upload services

    **Authentication**: JWT Bearer tokens required (except health endpoints)

    **Authorization**: Role-based access control (RBAC) with policies:
    - Customer: View own invoices
    - Employee: Create and edit draft invoices
    - Manager: Finalize and cancel invoices
    - Admin: Full access including audit logs

    **Base Path**: `/invoices/v1`

    **Scalar UI**: Available at `/invoices/scalar/v1` (development only)
  contact:
    name: MALIEV Co. Ltd. API Team
    email: api@maliev.com
  license:
    name: Proprietary
servers:
  - url: https://dev.api.maliev.com
    description: Development environment
  - url: https://staging.api.maliev.com
    description: Staging environment
  - url: https://api.maliev.com
    description: Production environment
security:
  - BearerAuth: []
...
paths:
...
  /invoices/v1/audit/invoices/{id}:
    get:
      tags:
        - Audit
      summary: Get audit trail for invoice
      description: |
        Retrieve complete audit trail for a specific invoice.
        Includes all lifecycle events with timestamps and actor information.
      operationId: getInvoiceAuditTrail
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/InvoiceIdParam'
      responses:
        '200':
          description: Audit trail retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoiceId:
                    type: string
                    format: uuid
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLogResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ========================================
  # Health Check Endpoints
  # ========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication.
        Required claims: NameIdentifier, Name, Email, Role, userType
  parameters:
    InvoiceIdParam:
      name: id
      in: path
      required: true
      description: Invoice UUID
      schema:
        type: string
        format: uuid
        example: 123e4567-e89b-12d3-a456-426614174000
...
  schemas:
...
    AuditLogResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        invoiceId:
          type: string
          format: uuid
        eventType:
          type: string
          enum: [Created, Updated, Finalized, Cancelled, PaymentLinked, Split]
        timestamp:
          type: string
          format: date-time
        actorId:
          type: string
          description: User ID or system identifier
        changedFields:
          type: object
          nullable: true
          description: JSON object of changed fields (for Update events)
        reason:
          type: string
          nullable: true
          description: Reason for action (e.g., cancellation reason)

    # ========================================
    # Error Schemas
    # ========================================
    ErrorResponse:
      type: object
      properties:
        type:
          type: string
          description: Error type URI
          example: https://tools.ietf.org/html/rfc7231#section-6.5.1
        title:
          type: string
          description: Short error description
          example: Bad Request
        status:
          type: integer
          description: HTTP status code
          example: 400
        detail:
          type: string
          description: Detailed error message
          example: Customer Tax ID must be 13 digits
        instance:
          type: string
          description: Request URI
          example: /invoices/v1/invoices
        traceId:
          type: string
          description: Correlation ID for tracing
          example: 0HN7CVQR3QVJT:00000001
        errors:
          type: object
          nullable: true
          description: Validation errors grouped by field
          additionalProperties:
            type: array
            items:
              type: string
  responses:
...
    Unauthorized:
      description: Unauthorized (missing or invalid JWT token)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden (insufficient permissions)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'